use existing sniproxy.conf file

+++ files/start-sni.sh
@@ -5,8 +5,14 @@ apt-get install autotools-dev cdbs debhelper dh-autoreconf dpkg-dev gettext libe
 echo "Installing sni proxy..."
 dpkg -i /server/sniproxy_0.6.1+git.5.g7fdd86c_amd64.deb
 
-echo "Copying sniproxy.conf..."
-cp /server/sniproxy.conf /etc/sniproxy.conf
+echo "Checking for existing sniproxy.conf..."
+if [ -f /data/sniproxy.conf ]; then
+    echo "sniproxy.conf already exists, using existing one..."
+    cp /data/sniproxy.conf /etc/sniproxy.conf
+else
+    echo "sniproxy.conf not found, copying default..."
+    cp /server/sniproxy.conf /etc/sniproxy.conf
+fi
 
 echo "Starting sniproxy..."
 sniproxy
